添加WorkThreadAI测试。

+++ purenessscopeserver/purenessscopeserver/PurenessScopeServer/Message/MessageService.cpp
@@ -251,7 +251,7 @@ bool CMessageService::ProcessMessage(CMessage* pMessage, uint32 u4ThreadID)
         m_ThreadInfo.m_u4CurrPacketCount++;
         m_ThreadInfo.m_u2CommandID   = u2CommandID;
 
-        bool blIsDead = m_WorkThreadAI.CheckCurrTimeout(pMessage->GetMessageBase()->m_u2Cmd, (uint32)m_ThreadInfo.m_tvUpdateTime.sec());
+        bool blIsDead = m_WorkThreadAI.CheckCurrTimeout(pMessage->GetMessageBase()->m_u2Cmd, (uint64)m_ThreadInfo.m_tvUpdateTime.sec());
 
         if(blIsDead == true)
         {
